requirements.txt
requests
textwrap3 ; python_version < "3.11"
python-dotenv
